+++ packages/components/rhf/src/index.ts
@@ -1,15 +1,16 @@
 "use client";
 
+export { RHFAutocomplete } from "./autocomplete";
 export { RHFCheckbox } from "./checkbox";
 export { RHFEditor } from "./editor";
-export { RHFInput } from "./input";
 export { RHFFileUploadMulti } from "./file-upload-multi";
+export { RHFFileUploadSingle } from "./file-upload-single";
+export { RHFInput } from "./input";
 export { RHFOtpInput } from "./otp-input";
 export { RHFProvider, type RHFProviderProps } from "./provider";
 export { RHFRadioGroup } from "./radio-group";
 export { RHFRating } from "./rating";
 export { RHFSelect } from "./select";
-export { RHFFileUploadSingle } from "./file-upload-single";
 export { RHFSwitch } from "./switch";
 export { RHFTagsInput } from "./tags-input";
 export { RHFTextarea } from "./textarea";