+++ apps/web/CHANGELOG.md
@@ -1,5 +1,12 @@
 # web
 
+## 0.1.165
+
+### Patch Changes
+
+- rhf autocomplete
+  - @jamsr-ui/react@0.0.183
+
 ## 0.1.164
 
 ### Patch Changes